# api/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Go Queueing System
  description: An API for Queueing System
  termsOfService: https://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: https://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://go-smo.aaletov.ru/api
paths:
  /source/getGenerated:
    get:
      summary: Get all requests generated by source
      operationId: GetAllGenRequests
      parameters:
        - name: sourceNum
          in: query
          required: true
          schema:
            type: integer
          allowReserved: true
      responses:
        '200':
          description: Get generated requests
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/ReqWT'
        '501':
          description: Generic error
  /source/getAll:
    get:
      summary: Get array of all source numbers
      operationId: GetAllSources
      responses:
        '200':
          description: Get all sourceNums
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/APISource'
  /buffer/getAll:
    get:
      summary: Get array of all buffer numbers
      operationId: GetAllBuffers
      responses:
        '200':
          description: Get all bufNums
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/APIBuffer'
  /buffer/getAllProcessed:
    get:
      summary: Get all requests processed by buffer
      operationId: GetAllBufProcessedRequests
      parameters:
        - name: bufNum
          in: query
          required: true
          schema:
            type: integer
          allowReserved: true
      responses:
        '200':
          description: Get processed
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/ReqWSE'
  /device/getAll:
    get:
      summary: Get array of all device numbers
      operationId: GetAllDevices
      responses:
        '200':
          description: Get all devNums
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/APIDevice'
  /device/getDone:
    get:
      summary: Get all requests done by device
      operationId: GetDeviceDoneRequests
      parameters:
        - name: devNum
          in: query
          required: true
          schema:
            type: integer
          allowReserved: true
      responses:
        '200':
          description: Get generated requests
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/ReqWT'
  /request/getAllRejected:
    get:
      summary: Get all rejected requests
      operationId: GetAllRejectedRequests
      responses:
        '200':
          description: Get all rejected requests
          content:
            application/json:
                schema:
                  type: array
                  items: 
                    $ref: '#/components/schemas/ReqWSE'
  /request/getAllDone:
    get:
      summary: Get all done requests
      operationId: GetAllDoneRequests
      responses:
        '200':
          description: Get all done requests
          content:
            application/json:
                schema:
                  type: array
                  items: 
                    $ref: '#/components/schemas/ReqWT'
components:
  schemas:
    ReqWT:
      type: object
      properties:
        request: 
          $ref: '#/components/schemas/Request'
        time:
          type: string
          format: date-time
    Request:
      type: object
      properties:
        sourceNumber:
          type: integer
        requestNumber:
          type: integer
    Buffer:
      type: object
      properties:
        bufNumber: 
          type: integer
        currentRequest:
          $ref: '#/components/schemas/Request'
    ReqWSE:
      type: object
      properties:
        request: 
          $ref: '#/components/schemas/Request'
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
    APISource:
      type: object
      properties:
        sourceNum:
          type: integer
    APIBuffer:
      type: object
      properties:
        bufNum:
          type: integer
    APIDevice:
      type: object
      properties:
        devNum:
          type: integer
        

        

        